Refactored the Search Profiles to be a separate Django application
