+++ listings/urls.py
@@ -1,7 +1,7 @@
 
 from django.urls import path
 
-from listings import views
+from . import views
 
 
 urlpatterns = [
