+++ templates/includes/users/search_profile.html
@@ -2,9 +2,9 @@
     {{ sp.title }} - {{ sp.city }} - {{ sp.listing_type }} - {{ sp.min_rooms }}
     - {{ sp.min_bedrooms }} - Price range: {{ sp.min_price }}: {{ sp.max_price }}
     - {{ sp.interior }} - {{ sp.frequency }} - {{ sp.is_active }}
-    - <a href="{% url 'accounts:sp_update' sp.pk %}">Update</a>
-    - <a href="{% url 'accounts:sp_delete' sp.pk %}"
+    - <a href="{% url 'sp:sp_update' sp.pk %}">Update</a>
+    - <a href="{% url 'sp:sp_delete' sp.pk %}"
          onclick="return confirm('Are you sure you want to delete this listing?')">Delete
       </a>
-    - <a href="{% url 'accounts:sp_toggle' sp.pk %}">{% if sp.is_active %}Disable{% else %}Enable{% endif %}</a>
+    - <a href="{% url 'sp:sp_toggle' sp.pk %}">{% if sp.is_active %}Disable{% else %}Enable{% endif %}</a>
 </p>
