+++ users/managers.py
@@ -1,4 +1,3 @@
-import uuid
 from django.contrib.auth.models import BaseUserManager
 from django.utils.translation import ugettext_lazy as _
